+++ src/main/java/com/dute/officialNetwork/domain/repository/DecorationClassRepository.java
@@ -10,5 +10,5 @@ import com.dute.officialNetwork.domain.entity.DecorationClass;
 public interface DecorationClassRepository extends JpaRepository<DecorationClass, Long> {
 
 	@Query(value = "SELECT decoration_class.* FROM decoration_class LEFT JOIN decoration_class_sub_type ON dcs_id = dc_dcs_id LEFT JOIN decoration_class_main_type ON dcm_id = dcs_dcm_id WHERE dcm_id = ?1 ORDER BY dc_create_time DESC LIMIT 0,7;", nativeQuery = true)
-	List<DecorationClass> findByMainTypeIdOrderByCreateTime(Long id);
+	List<DecorationClass> findByMainTypeIdOrderByCreateTime(Integer id);
 }
